[abi] Improved codecs, deflate compression & decompression code
  * Version bump to 0.12.0

  * Added a new DEFLATE decompressor that is much faster than the
    previous implementation. It has some unique optimizations for
    decoding up to 2 literals per decode entry
  * Fixed a bug in a DEFLATE compressor (long litlen) - this was
    not hit in practice, but specially crafted inputs compressed
    wrongly
  * Fixed a bug in a BMP decoder validation of image data offset
  * Added support for decoding APNG images (animated PNGs)
  * Reorganized some code in a DEFLATE compressor, especially code
    that is writing the output bit-stream, to make it a little bit
    faster
  * Added more DEFLATE tests (special cases, round-trip encoding &
    decoding)
  * Added more image codec tests (round-trip encoding + decoding)
  * Added an optimized CRC32 checksum computations that use PCLMULQDQ
    (X86) and ASIMD+CRYPTO (ARM) extensions
  * Added an optimized ADLER32 checksum computations that use SSE2
    (X86) and ASIMD (ARM) extensions
  * Added '-Wconversion' warning to always be used when compiling
    Blend2D (and made the code clean to compile with this warning)
  * Improved all codecs to fill more information in BLImageInfo
  * Improved bl::ArenaAllocator to use dynamic block sizes when
    allocating more than 2 block

+++ src/blend2d/image.cpp
@@ -55,7 +55,7 @@ static BL_INLINE BLResultT<intptr_t> calcStrideFromCreateParams(int w, int h, BL
     return BLResultT<intptr_t>{result, 0};
   }
   else {
-    uint32_t bytesPerLine = strideForWidth(w, blFormatInfo[format].depth);
+    uint32_t bytesPerLine = strideForWidth(uint32_t(w), blFormatInfo[format].depth);
     uint64_t bytesPerImage = uint64_t(bytesPerLine) * unsigned(h);
 
     // NOTE: Align the stride to 16 bytes if bytesPerLine is not too small. The reason is that when multi-threaded
@@ -71,7 +71,7 @@ static BL_INLINE BLResultT<intptr_t> calcStrideFromCreateParams(int w, int h, BL
 
 // Make sure that the external image won't cause any kind of overflow in rasterization and texture fetching.
 static BL_INLINE BLResult checkCreateFromDataParams(int w, int h, BLFormat format, intptr_t stride) noexcept {
-  uint32_t minimumStride = strideForWidth(w, blFormatInfo[format].depth);
+  uint32_t minimumStride = strideForWidth(uint32_t(w), blFormatInfo[format].depth);
   uintptr_t bytesPerLine = uintptr_t(blAbs(stride));
 
   if (BL_UNLIKELY(!checkSizeAndFormat(w, h, format) || bytesPerLine < minimumStride))
@@ -393,7 +393,8 @@ BL_API_IMPL BLResult blImageMakeMutable(BLImageCore* self, BLImageData* dataOut)
 
   if (format != BL_FORMAT_NONE && !isImplMutable(selfI)) {
     BLImageCore newO;
-    BL_PROPAGATE(allocImpl(&newO, size.w, size.h, format, strideForWidth(size.w, blFormatInfo[format].depth)));
+    BL_PROPAGATE(allocImpl(&newO, size.w, size.h, format,
+      intptr_t(strideForWidth(uint32_t(size.w), blFormatInfo[format].depth))));
 
     BLImagePrivateImpl* newI = getImpl(&newO);
     dataOut->pixelData = newI->pixelData;
